fix ingestion tests
